+++ scripts/build-css.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const fs = require('fs');
-const bundleCss = require('./utils/bundle-css');
+const bundleCss = require('@discoveryjs/cli/lib/shared/bundle-css');
 const inputFilename = path.join(__dirname, '../client/lib.css');
 const outputFilename = path.join(__dirname, '../dist/lib.css');