+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 ## next
 
-- Removed `src/scan-fs.js` module since it extracted into a separate package – [@discoveryjs/scan-fs](https://github.com/discoveryjs/scan-fs)
-- Server
-    - Fixed asset generation in model-free mode
+- Removed `src/scan-fs.js` module since extracted into a separate package [@discoveryjs/scan-fs](https://github.com/discoveryjs/scan-fs)
+- Extracted CLI tools to separate package [@discoveryjs/cli](https://github.com/discoveryjs/discovery-cli)
 
 ## 1.0.0-beta.17 (06-09-2019)